requirements.txt
# Core RL and ML libraries
gymnasium==0.29.1
numpy>=1.26.0
pandas>=2.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Deep Learning (for DQN)
torch>=2.0.0
tensorboard>=2.13.0

# Dashboard and API
fastapi>=0.100.0
uvicorn>=0.23.0
websockets>=11.0.0
python-socketio>=5.8.0

# Data visualization
plotly>=5.15.0
dash>=2.11.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
requests>=2.31.0

# Development
jupyter>=1.0.0
ipykernel>=6.25.0 